Move login and camera screens into stack navigators
