+++ package.json
@@ -43,6 +43,7 @@
     "@react-navigation/drawer": "^6.6.2",
     "@react-navigation/elements": "^1.3.17",
     "@react-navigation/native": "^6.1.6",
+    "@react-navigation/native-stack": "^6.9.13",
     "@react-spring/native": "9.5.5",
     "@realm/react": "^0.4.3",
     "@sayem314/react-native-keep-awake": "^1.1.0",
